requirements.txt
strands-agents
requests
rich
langchain_aws
langgraph
python-dotenv